Fix RC & RD.
The naming convention I'm going for is that signals are named by
the computers that this glue is binding. So signals from the TI start
with ti_, and the name of the signal is then the name the TI gives it.
For example ti_cru_out. It is an input to the glue logic, but it is
and _out from the TI.

So, now following this, I've fixed the signal names for the Raspberry PI.
rpi_  _out, _in. in particular for the shift register sdata signals.