fix unicode handling in linux file names
Now if there is any unicode or the name is longer than 10 characters,
the full name is processed by the CRC15 to come up with the hash suffix.
the full name is transliterated to approximate single byte characters
for any unicode character, and then the prefix is take from that.