rudimentory json parsing via file interface
open a file with the ?J native format override, such as:
  `TIPI.?J.JSONDATA` or `PI.?J.https://jokeoftheday.org`
and TIPI will return a dynamic file that can be queried.
write a record with a JMESPath query string, and the record
set for the dynamic file will be reset to just that of the
results of the query.

Still needs some TI-99 cleanup