Pack native .fb files into Forth blocks
rough draft, seems to work.

This is a read-only operation. Native text files ending in the `.fb` suffix
can be loaded as a packed set of blocks into FbForth. Should work with other
DISPLAY/FIXED 128 type forth variants on the 4A as well, such as TurboForth.

```
USEBFL TIPI.4TH/FB
1 LOAD
```

As blocks fill up, the block continue command is added, and the next block
is added and packed.